[feat]room, loginのurl修正
