+++ app/front/pages/home.vue
@@ -10,7 +10,7 @@
       <div class="home-top__event--title">開催中のイベント</div>
       <div v-for="(room, id) in ongoingRooms" :key="id">
         <NuxtLink
-          :to="{ path: '/', query: { roomId: room.id, user: 'admin' } }"
+          :to="'/room/' + room.id + '/?user=admin'"
           class="home-top__event--item"
         >
           <div class="home-top__event__list">
@@ -27,7 +27,7 @@
       <div class="home-top__event--title">未開始のイベント</div>
       <div v-for="(room, id) in notStartedRooms" :key="id">
         <NuxtLink
-          :to="{ path: '/', query: { roomId: room.id, user: 'admin' } }"
+          :to="'/room/' + room.id + '/?user=admin'"
           class="home-top__event--item"
         >
           <div class="home-top__event__list">
@@ -42,7 +42,7 @@
       <div class="home-top__event--title">終了済みのイベント</div>
       <div v-for="(room, id) in finishedRooms" :key="id">
         <NuxtLink
-          :to="{ path: '/', query: { roomId: room.id, user: 'admin' } }"
+          :to="'/room/' + room.id + '/?user=admin'"
           class="home-top__event--item"
         >
           <div class="home-top__event__list">
@@ -160,7 +160,7 @@ export default Vue.extend({
     async logout() {
       try {
         await this.$fire.auth.signOut()
-        this.$router.push("/login")
+        this.$router.push("/")
       } catch {
         alert("ログアウトに失敗しました")
       }
